migrate charm libs and deprecate old ones (#371)
* migrated charm libs and deprecated old ones

* fix renamee in parca libs
